+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [1.1.0] 2017-02-01
+### Additions
+* cowsay greeting is the new default greeting
+
+### Changes
+* plain greeting is preserved with `--plain` flag
+
 ## [1.0.0] 2017-01-25
 ### Additions
 * enforces coding standards in .eslintrc